+++ asdf/__init__.py
@@ -19,14 +19,13 @@ __all__ = [
 ]
 
 
-from jsonschema import ValidationError
-
 from ._convenience import info
 from ._types import CustomType
 from ._version import version as __version__
 from .asdf import AsdfFile
 from .asdf import open_asdf as open
 from .config import config_context, get_config
+from .exceptions import ValidationError
 from .stream import Stream
 from .tags.core import IntegerType
 from .tags.core.external_reference import ExternalArrayReference