+++ asdf/_tests/tags/core/tests/test_ndarray.py
@@ -3,7 +3,6 @@ import os
 import re
 import sys
 
-import jsonschema
 import numpy as np
 import pytest
 import yaml
@@ -14,7 +13,7 @@ import asdf
 from asdf import util
 from asdf._tests import _helpers as helpers
 from asdf._tests.objects import CustomTestType
-from asdf.exceptions import AsdfDeprecationWarning
+from asdf.exceptions import AsdfDeprecationWarning, ValidationError
 from asdf.tags.core import ndarray
 
 from . import data as test_data
@@ -538,7 +537,7 @@ def test_invalid_mask_datatype(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r".* is not valid under any of the given schemas"), asdf.open(
+    with pytest.raises(ValidationError, match=r".* is not valid under any of the given schemas"), asdf.open(
         buff,
     ):
         pass
@@ -552,7 +551,7 @@ def test_ndim_validation(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r"Wrong number of dimensions:.*"), asdf.open(
+    with pytest.raises(ValidationError, match=r"Wrong number of dimensions:.*"), asdf.open(
         buff,
         extensions=CustomExtension(),
     ):
@@ -606,7 +605,7 @@ def test_ndim_validation(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r"Wrong number of dimensions:.*"), asdf.open(
+    with pytest.raises(ValidationError, match=r"Wrong number of dimensions:.*"), asdf.open(
         buff,
         extensions=CustomExtension(),
     ):
@@ -633,7 +632,7 @@ def test_datatype_validation(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r"Can not safely cast from .* to .*"), asdf.open(
+    with pytest.raises(ValidationError, match=r"Can not safely cast from .* to .*"), asdf.open(
         buff,
         extensions=CustomExtension(),
     ):
@@ -658,7 +657,7 @@ def test_datatype_validation(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r"Expected datatype .*, got .*"), asdf.open(
+    with pytest.raises(ValidationError, match=r"Expected datatype .*, got .*"), asdf.open(
         buff,
         extensions=CustomExtension(),
     ):
@@ -676,7 +675,7 @@ def test_datatype_validation(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r"Expected scalar datatype .*, got .*"), asdf.open(
+    with pytest.raises(ValidationError, match=r"Expected scalar datatype .*, got .*"), asdf.open(
         buff,
         extensions=CustomExtension(),
     ):
@@ -711,7 +710,7 @@ def test_structured_datatype_validation(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r"Can not safely cast to expected datatype.*"), asdf.open(
+    with pytest.raises(ValidationError, match=r"Can not safely cast to expected datatype.*"), asdf.open(
         buff,
         extensions=CustomExtension(),
     ):
@@ -731,7 +730,7 @@ def test_structured_datatype_validation(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r"Mismatch in number of columns:.*"), asdf.open(
+    with pytest.raises(ValidationError, match=r"Mismatch in number of columns:.*"), asdf.open(
         buff,
         extensions=CustomExtension(),
     ):
@@ -744,7 +743,7 @@ def test_structured_datatype_validation(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r"Expected structured datatype.*"), asdf.open(
+    with pytest.raises(ValidationError, match=r"Expected structured datatype.*"), asdf.open(
         buff,
         extensions=CustomExtension(),
     ):
@@ -762,7 +761,7 @@ def test_structured_datatype_validation(tmpdir):
     """
     buff = helpers.yaml_to_asdf(content)
 
-    with pytest.raises(jsonschema.ValidationError, match=r"Expected datatype .*, got .*"), asdf.open(
+    with pytest.raises(ValidationError, match=r"Expected datatype .*, got .*"), asdf.open(
         buff,
         extensions=CustomExtension(),
     ):