+++ CHANGES.rst
@@ -5,6 +5,7 @@ The ASDF Standard is at v1.6.0
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 - Drop Python 3.8 support [#1556]
+- Drop NumPy 1.20, 1.21 support [#1568]
 
 
 2.15.0 (2023-03-28)