+++ CHANGES.rst
@@ -8,6 +8,7 @@ The ASDF Standard is at v1.6.0
 - Drop NumPy 1.20, 1.21 support [#1568]
 - Convert numpy scalars to python types during yaml encoding
   to handle NEP51 changes for numpy 2.0 [#1605]
+- Vendorize jsonschema 4.17.3 [#1591]
 
 2.15.0 (2023-03-28)
 -------------------